# internal/app/openapi.yaml
openapi: 3.1.0
info:
  title: Kore API
  version: 1.0.0
  description: API Kore — monolithe modulaire PSA/ESN
servers:
  - url: http://localhost:8080/api/v1
paths:
...
  /integrations/connections:
    get:
      summary: Liste connexions intégration
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: OK }
    post:
      summary: Créer connexion intégration
      security: [{ bearerAuth: [] }]
      responses:
        "201": { description: Créé }
  /integrations/connections/{id}:
    get:
      summary: Détail connexion
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "200": { description: OK }
    delete:
      summary: Supprimer connexion
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "200": { description: OK }
  /integrations/connections/{id}/sync:
    post:
      summary: Synchroniser connexion
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "200": { description: OK }
  /integrations/api-keys:
    get:
      summary: Liste clés API
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: OK }
    post:
      summary: Créer clé API
      security: [{ bearerAuth: [] }]
      responses:
        "201": { description: Créé }
  /integrations/api-keys/{id}:
    delete:
      summary: Révoquer clé API
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "200": { description: OK }
...
  /admin/parameters:
    get:
      summary: Paramètres admin tenant
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: OK }
    put:
      summary: Mettre à jour paramètres admin
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: OK }
  /admin/templates:
    get:
      summary: Liste modèles admin
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: OK }
    post:
      summary: Créer modèle admin
      security: [{ bearerAuth: [] }]
      responses:
        "201": { description: Créé }
  /admin/templates/{id}:
    get:
      summary: Détail modèle
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "200": { description: OK }
  /admin/phone-directory:
    get:
      summary: Annuaire téléphonique
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: OK }
    post:
      summary: Ajouter entrée annuaire
      security: [{ bearerAuth: [] }]
      responses:
        "201": { description: Créé }
...
  /ett/clock-in:
    post:
      summary: Pointage entrée ETT
      security: [{ bearerAuth: [] }]
      responses:
        "201": { description: Créé }
  /ett/clock-out:
    post:
      summary: Pointage sortie ETT
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: OK }
  /ett/records:
    get:
      summary: Relevés ETT
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: OK }
  /ett/records/{id}/correct:
    post:
      summary: Correction tracée ETT
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "200": { description: OK }
  /ett/records/{id}/audit:
    get:
      summary: Journal audit ETT
      security: [{ bearerAuth: [] }]
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "200": { description: OK }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT